Refactoring:  names and ServiceLayer className
