+++ components/home/YoutubeSection.tsx
@@ -1,4 +1,4 @@
-import { ContentsDataType } from '@/models/sheet/Insheet';
+import { ContentsDataType } from '@/models/sheet/InSheet';
 import home from '@/styles/home/Home.module.scss';
 import YoutubeContentCard from '@/components/YoutubeContentCard';
