+++ hooks/UseScheduledData.ts
@@ -1,9 +1,9 @@
-import { sheetAPIReturnType } from '@/models/sheet/Insheet';
+import { SheetAPIReturnType } from '@/models/sheet/InSheet';
 import useSWR from 'swr';
 import { fetcher } from '@/hooks/Fetcher';
 
 const useScheduledData = () => {
-  const { data, error, mutate, isLoading } = useSWR('/api/sheet/upcoming', fetcher<sheetAPIReturnType>(), {
+  const { data, error, mutate, isLoading } = useSWR('/api/sheet/upcoming', fetcher<SheetAPIReturnType>(), {
     dedupingInterval: 15 * 1000,
     errorRetryCount: 3,
     errorRetryInterval: 3 * 1000,
