+++ pages/api/sheet/all.ts
@@ -1,12 +1,12 @@
 import type { NextApiRequest, NextApiResponse } from 'next';
-import { sheetAPIReturnType } from '@/models/sheet/Insheet';
+import { SheetAPIReturnType } from '@/models/sheet/InSheet';
 import { getNow } from '@/utils/GetTime';
 import { getSheet } from '@/models/sheet/Sheets';
 import parseYoutubeContentData from '@/utils/ParseSheetData';
 import getENV from '@/utils/GetENV';
-import { CONTENTS_SHEET_ID, CONTENTS_SHEET_RANGE } from '@/const';
+import { CONTENTS_SHEET_ID, CONTENTS_SHEET_RANGE } from '@/consts';
 
-const handler = async (req: NextApiRequest, res: NextApiResponse<sheetAPIReturnType>) => {
+const handler = async (req: NextApiRequest, res: NextApiResponse<SheetAPIReturnType>) => {
   try {
     if (req.method !== 'GET') throw new Error('invaild method');
     const spreadsheetId = getENV(CONTENTS_SHEET_ID);
